requirements.txt
Flask
easyocr
torch
torchvision
Pillow
python-dotenv
pdf2image
numpy